Move debugInfo to a separate module
