+++ karma.conf.js
@@ -13,6 +13,7 @@ module.exports = function(config) {
         'test/vendor/json2.js',
         'test/vendor/underscore.js',
         'backbone.js',
+        'debug-info.js',
         'test/setup/*.js',
         'test/*.js'
     ],
