+++ test/index.html
@@ -11,6 +11,7 @@
   <script src="vendor/json2.js"></script>
   <script src="vendor/underscore.js"></script>
   <script src="../backbone.js"></script>
+  <script src="../debug-info.js"></script>
   <script src="setup/dom-setup.js"></script>
   <script src="setup/environment.js"></script>
   <script src="noconflict.js"></script>
